improvement made to the Cosmos client and created separate plugin class for setting priority header
